+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Invalidate energy capabilities on removed tile entities [#3]
+
 ## [0.3.3] - 2019-08-17
 ### Fixed
 - Server crash [#1]